Refactors components into separate modules
Extracts inline components from the admin user route into dedicated component files to improve code organization and reusability.

Creates separate modules for user profile display, account statistics, activity filtering, and activity tables. Also introduces a reusable StateButton component for edit/save/cancel functionality.

Improves activity sorting to display most recent entries first and adds proper styling utilities to the component style system.
